==> 01_data_preparation.ipynb <==
Data to prepare:
1. Dataset-publication Interaction (Graph DP)
2. Dataset features
3. Publication features

PATH_data_import = "/Users/lizhouf/Documents/4_MICA/Scholarly Object Management + RecSys/DataChat/Data_raw/"
PATH_data_export = "/Users/lizhouf/Documents/4_MICA/Scholarly Object Management + RecSys/DataChat/Data_result/"

import pandas as pd
import numpy as np
import re
import json

df_studies_all = pd.read_csv(PATH_data_import+"ICPSR_datasets.csv")
df_studies_meta = pd.read_csv(PATH_data_import+"study_data_20230217.csv")
df_bib = pd.read_csv(PATH_data_import+"ICPSR_publications.csv")
df_stud_bib = pd.read_csv(PATH_data_import+"ICPSR_datasets_publications_interaction.csv")

# Preprocess datasets

df_studies_all.fillna("NoRecord",inplace=True)

df_studies_all["DESCRIPTION"]=df_studies_all["DESCRIPTION_1"]+df_studies_all["DESCRIPTION_2"]+df_studies_all["DESCRIPTION_3"]+df_studies_all["DESCRIPTION_4"]+df_studies_all["DESCRIPTION_5"]
def striphtml(data):
    p = re.compile(r'<.*?>')
    return p.sub('', str(data))
df_studies_all.DESCRIPTION = df_studies_all.DESCRIPTION.apply(lambda x: striphtml(x))

df_studies_all.drop(columns = ['ALTTITLE1', 'ALTTITLE2', 'ALTTITLE3', 'ALTTITLE4',
                               'ALTTITLE5', 'ALTTITLE6', 'ALTTITLE7',
                               'DESCRIPTION_1', 'DESCRIPTION_2','DESCRIPTION_3',
                               'DESCRIPTION_4', 'DESCRIPTION_5'],inplace=True)

df_studies_all.STUDY = df_studies_all.STUDY.astype(int).astype(int)

df_studies_all.ORIGRELDATE = df_studies_all.ORIGRELDATE.apply(lambda x: x[:10])

df_studies_all["YEAR"] = df_studies_all.ORIGRELDATE.apply(lambda x: x[:4])

# limit to datasets after 2017
df_studies_all = df_studies_all[df_studies_all.YEAR.isin(["2017","2018","2019","2020","2021","2022"])].reset_index(drop=True)

df_studies_all.info()

# augment 1: add more metadata

users_dict = dict(zip(df_studies_meta.STUDY,df_studies_meta[" users_to_20220822"]))
data_users_dict = dict(zip(df_studies_meta.STUDY,df_studies_meta[" datausers_to_20220822"]))
df_studies_all["users"] = [0]*len(df_studies_all)
df_studies_all["data_users"] = [0]*len(df_studies_all)
for i in range(len(df_studies_all)):
    this_study = df_studies_all.STUDY[i]
    if this_study in users_dict:
        this_num = users_dict[this_study]
        try:
            df_studies_all["users"][i] = int(this_num)
        except:
            0
    if this_study in data_users_dict:
        this_num = data_users_dict[this_study]
        try:
            df_studies_all["data_users"][i] = int(this_num)
        except:
            0


# augment 2: add data citing number: NOT NEEDED, REMOVE THIS!

df_studies_all["citations"] = [0]*len(df_studies_all)
for data_refs in df_bib.STUD_NUMS:
    try:
        data_refs_list = data_refs.split(";")
        for ref in data_refs_list:
            try: 
                df_studies_all.loc[df_studies_all.STUDY==int(ref),"citations"]+=1
            except:
                0
    except:
        0

df_bib.columns

# augment 3: add publications details

def remove_duplicate_dict(l):
    seen = set()
    new_l = []
    for d in l:
        t = d["name"]
        if t not in seen:
            seen.add(t)
            new_l.append(d)
    return new_l

 remove_duplicate_dict([{"name":"a","pubRefCount":1},{"name":"a","pubRefCount":3},{"name":"b","pubRefCount":3}])

# name, url, citations
publication_list = []

for i in range(len(df_studies_all)):
    if i%100==0:
        print(i)
    this_study = df_studies_all.STUDY[i]
    this_pubs_list = []
    
    for j in range(len(df_bib)):
        data_refs = df_bib.STUD_NUMS[j]
        if type(data_refs)==str:
            data_refs_list = [int(ref) for ref in data_refs.split(";")]
            if this_study in data_refs_list:
                ref_count = df_bib.times_cited[j]
                if np.isnan(ref_count):
                    ref_count = 0
                this_pubs_list.append({"name": df_bib.TITLE[j],
                                       "pubRefCount": ref_count})

    publication_list.append(remove_duplicate_dict(this_pubs_list))

df_studies_all["publications"] = publication_list

# export results

data_results = list()
for idx, row in df_studies_all.iterrows():
    data_results.append({"id":row["STUDY"],
                         "name":row["NAME"],
                         "date":row["ORIGRELDATE"],
                         "url":"https://doi.org/"+row["DOI"], 
                         "series":row["SERIES_TITLE"],
                         "owner":row["OWNER"],
                         "users":row["users"],
                         "dataUsers":row["data_users"],
                         "dataReferences":row["citations"],
                         "funders":[{"name":funder} for funder in row["FUNDINGAGENCY"].split(".")], 
                         "locations":[{"name":location} for location in row["GEO"].split(";")],
                         "terms":[{"name":term} for term in row['TERMS'].split(";")],
                         "publications":row["publications"]
                        })

data_results[:3]

len(data_results)

with open(PATH_data_export+"data_output_17_22.json", "w") as fout:
    json.dump(data_results, fout)
